requirements.txt
uvicorn
fastapi
pydantic
flask==2.3.3
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.28.1
pandas==2.1.3
undetected-chromedriver==3.5.4
